first day off

+++ DatsartSpace/API/Models/Shot.cs
@@ -2,8 +2,8 @@
 
 public class Shot
 {
-    public float AngleHorizontal { get; set; }
-    public float AngleVertical { get; set; }
+    public double AngleHorizontal { get; set; }
+    public double AngleVertical { get; set; }
     public object ColorPot { get; set; }
-    public float Power { get; set; }
+    public double Power { get; set; }
 }
